Eliminate controller duplication (pretty trace tab)
