+++ apps/block_scout_web/lib/block_scout_web/templates/layout/app.html.eex
@@ -74,8 +74,7 @@
       @view_module != Elixir.BlockScoutWeb.PendingTransactionView &&
       @view_module != Elixir.BlockScoutWeb.TransactionInternalTransactionView &&
       @view_module != Elixir.BlockScoutWeb.TransactionLogView &&
-      @view_module != Elixir.BlockScoutWeb.TransactionRawTraceView &&
-      @view_module != Elixir.BlockScoutWeb.TransactionPrettyTraceView &&
+      @view_module != Elixir.BlockScoutWeb.TransactionTraceView &&
       @view_module != Elixir.BlockScoutWeb.TransactionTokenTransferView &&
       @view_module != Elixir.BlockScoutWeb.AddressTransactionView &&
       @view_module != Elixir.BlockScoutWeb.AddressTokenTransferView &&
