+++ apps/block_scout_web/lib/block_scout_web/templates/transaction/_tabs.html.eex
@@ -24,11 +24,11 @@
   <%= link(
       gettext("Raw Trace"),
       class: "card-tab #{tab_status("raw_trace", @conn.request_path)}",
-      to: transaction_raw_trace_path(@conn, :index, @transaction) <> "#tabs"
+      to: transaction_raw_trace_path(@conn, :raw, @transaction) <> "#tabs"
       ) %>
   <%= link(
       gettext("Pretty Trace"),
       class: "card-tab #{tab_status("pretty_trace", @conn.request_path)}",
-      to: transaction_pretty_trace_path(@conn, :index, @transaction) <> "#tabs"
+      to: transaction_pretty_trace_path(@conn, :pretty, @transaction) <> "#tabs"
       ) %>
 </div>
