--- apps/block_scout_web/lib/block_scout_web/templates/transaction_raw_trace/_metatags.html.eex
@@ -1,1 +1,0 @@
-<%= render BlockScoutWeb.TransactionView, "_metatags.html", conn: @conn, transaction: @transaction %>
