+++ apps/block_scout_web/lib/block_scout_web/templates/transaction_trace/_table_pretty_trace.html.eex
